requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchdata==0.7.0
torchtext==0.16.0+cpu
torchvision==0.16.0+cu118
matplotlib==3.7.3
seaborn==0.13.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.2.2
mlxtend==0.23.0
nltk==3.8.1
jupyterlab==4.0.6
tqdm==4.66.1
ipywidgets==8.1.1
bs4==0.0.1
lxml==4.9.3
wordcloud==1.9.2
Wikipedia-API==0.6.0
cssutils==2.7.1
eli5==0.13.0
lightning==2.1.0
trimesh[easy]=4.0.0
pysdf==0.1.9
PyMCubes==0.1.4
tensorboard==2.14.1
torch-scatter==2.1.2
scikit-image==0.22.0
